Move PublicationStatus to models.publication.
