+++ mobilizon_bots/event/event.py
@@ -3,17 +3,11 @@ from enum import IntEnum
 from typing import Optional
 
 import arrow
-from jinja2 import Template
 import tortoise.timezone
+from jinja2 import Template
 
 from mobilizon_bots.models.event import Event
-
-
-class PublicationStatus(IntEnum):
-    WAITING = 1
-    FAILED = 2
-    PARTIAL = 3
-    COMPLETED = 4
+from mobilizon_bots.models.publication import PublicationStatus
 
 
 class NotificationStatus(IntEnum):
