+++ tests/conftest.py
@@ -12,12 +12,11 @@ from mobilizon_bots.config.config import build_and_validate_settings
 
 from mobilizon_bots.event.event import (
     MobilizonEvent,
-    PublicationStatus,
     NotificationStatus,
 )
 from mobilizon_bots.models.event import Event
 from mobilizon_bots.models.notification import Notification
-from mobilizon_bots.models.publication import Publication
+from mobilizon_bots.models.publication import Publication, PublicationStatus
 from mobilizon_bots.models.publisher import Publisher
